+++ plugins/_browser_agent/extensions/python/_functions/agent/Agent/get_browser_model/start/_10_browser_agent.py
@@ -1,0 +1,7 @@
+from helpers.extension import Extension
+from plugins._browser_agent.helpers.browser_llm import build_browser_model_for_agent
+
+class BrowserModelProvider(Extension):
+    def execute(self, data: dict = {}, **kwargs):
+        if self.agent:
+            data["result"] = build_browser_model_for_agent(self.agent)
