--- plugins/_model_config/extensions/python/_functions/agent/Agent/get_browser_model/start/_10_model_config.py
@@ -1,8 +1,0 @@
-from helpers.extension import Extension
-from plugins._model_config.helpers.model_config import build_browser_model
-
-
-class BrowserModelProvider(Extension):
-    def execute(self, data: dict = {}, **kwargs):
-        if self.agent:
-            data["result"] = build_browser_model(self.agent)
