+++ plugins/_model_config/helpers/model_config.py
@@ -189,17 +189,6 @@ def build_utility_model(agent=None):
     )
 
 
-def build_browser_model(agent=None):
-    """Build and return the browser-use adapter using chat model config."""
-    cfg = get_chat_model_config(agent)
-    mc = build_model_config(cfg, models.ModelType.CHAT)
-    from plugins._browser_agent.helpers.browser_llm import (
-        build_browser_model_from_config,
-    )
-
-    return build_browser_model_from_config(mc)
-
-
 def build_embedding_model(agent=None):
     """Build and return an embedding model wrapper."""
     cfg = get_embedding_model_config(agent)
